修正实际项目中的问题。headView和FooterView 中不能做能改变高度的操作，比如对view的visibily设置GONE.因为layout中根据高度来scroll，如果改变高度，layout会重新布局，会出现跳一下的现象

+++ app/src/main/java/custom/sunday/com/sharedemo/component/refreshlayout/HeaderView.java
@@ -26,5 +26,5 @@ public interface HeaderView {
 
     boolean isPauseTime();
 
-    int getPauseMillTime();
+    long getPauseMillTime();
 }
